requirements.txt
# Bibliotecas para FastAPI e Uvicorn
fastapi==0.100.0
uvicorn==0.22.0
starlette==0.27.0  # Biblioteca necessária para CORS

# Bibliotecas do projeto
langchain==0.0.330
openai==0.28.1
python-dotenv==1.0.0
pydantic==2.4.2
pydantic_core==2.10.1
beautifulsoup4==4.12.2

# Leitura e extração de texto de PDFs
pdfplumber==0.10.0
PyPDF2==2.11.2

# Banco de dados vetorial e embeddings
faiss-cpu==1.7.4

# Processamento e manipulação de dados
numpy==1.26.1
pandas==2.1.2
tiktoken==0.5.1

# Outras dependências úteis
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.1
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
Pillow==10.1.0
protobuf==4.25.0
pyarrow==14.0.0
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
streamlit==1.28.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
yarl==1.9.2
youtube-transcript-api==0.6.1
zipp==3.17.0
